requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
python-multipart
